add LevelMgr

+++ Project/Engine/CConstBuffer.cpp
@@ -19,7 +19,7 @@ int CConstBuffer::Create(size_t _bufferSize, CB_TYPE _Type)
 	m_Type = _Type;
 
 	// 상수버퍼 생성
-	m_Desc.ByteWidth = _bufferSize;
+	m_Desc.ByteWidth = (UINT)_bufferSize;
 	m_Desc.MiscFlags = 0;
 	m_Desc.BindFlags = D3D11_BIND_CONSTANT_BUFFER;
